Update README formatting and content
Replace reference to SPIFBlockDevice as default, this example uses an
instance of the SDBlockDevice. Add note to explain what happens in the
scenario that a SD card is not found on the board. Update template to
follow format used in mbed-os-example-blinky. Remove linker table and
depracated compilers. Output in README matches output from source